// src/vendor/prettier-types.ts
export type VendoredPrettierOptions = Partial<VendoredPrettierOptionsRequired>;

export type VendoredPrettierRuleOptions = VendoredPrettierOptions & {
  parser?: BuiltInParserName | ExternalParserName;
  // eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
  [k: string]: unknown | undefined;
};

export interface VendoredPrettierOptionsRequired {
  /**
   * Specify the line length that the printer will wrap on.
   *
   * @default 120
   */
  printWidth: number;
  /**
   * Specify the number of spaces per indentation-level.
   */
  tabWidth: number;
  /**
   * Indent lines with tabs instead of spaces
   */
  useTabs?: boolean;
  /**
   * Print semicolons at the ends of statements.
   */
  semi: boolean;
  /**
   * Use single quotes instead of double quotes.
   */
  singleQuote: boolean;
  /**
   * Use single quotes in JSX.
   */
  jsxSingleQuote: boolean;
  /**
   * Print trailing commas wherever possible.
   */
  trailingComma: 'all' | 'es5' | 'none';
  /**
   * Print spaces between brackets in object literals.
   */
  bracketSpacing: boolean;
  /**
   * Put the `>` of a multi-line HTML (HTML, XML, JSX, Vue, Angular) element at the end of the last line instead of being
   * alone on the next line (does not apply to self closing elements).
   */
  bracketSameLine: boolean;
  /**
   * Put the `>` of a multi-line JSX element at the end of the last line instead of being alone on the next line.
   *
   * @deprecated use bracketSameLine instead
   */
  jsxBracketSameLine: boolean;
  /**
   * Format only a segment of a file.
   */
  rangeStart: number;
  /**
   * Format only a segment of a file.
   *
   * @default Number.POSITIVE_INFINITY
   */
  rangeEnd: number;
  /**
   * By default, Prettier will wrap markdown text as-is since some services use a linebreak-sensitive renderer.
   * In some cases you may want to rely on editor/viewer soft wrapping instead, so this option allows you to opt out.
   *
   * @default "preserve"
   */
  proseWrap: 'always' | 'never' | 'preserve';
  /**
   * Include parentheses around a sole arrow function parameter.
   *
   * @default "always"
   */
  arrowParens: 'always' | 'avoid';
  /**
   * Provide ability to support new languages to prettier.
   */
  plugins: (any | string)[];
  /**
   * How to handle whitespaces in HTML.
   *
   * @default "css"
   */
  htmlWhitespaceSensitivity: 'css' | 'ignore' | 'strict';
  /**
   * Which end of line characters to apply.
   *
   * @default "lf"
   */
  endOfLine: 'auto' | 'cr' | 'crlf' | 'lf';
  /**
   * Change when properties in objects are quoted.
   *
   * @default "as-needed"
   */
  quoteProps: 'as-needed' | 'consistent' | 'preserve';
  /**
   * Whether or not to indent the code inside <script> and <style> tags in Vue files.
   *
   * @default false
   */
  vueIndentScriptAndStyle: boolean;
  /**
   * Enforce single attribute per line in HTML, XML, Vue and JSX.
   *
   * @default false
   */
  singleAttributePerLine: boolean;

  /**
   * How to handle whitespaces in XML.
   *
   * @default "preserve"
   */
  xmlQuoteAttributes: 'double' | 'preserve' | 'single';
  /**
   * Whether to put a space inside the brackets of self-closing XML elements.
   *
   * @default true
   */
  xmlSelfClosingSpace: boolean;
  /**
   * Whether to sort attributes by key in XML elements.
   *
   * @default false
   */
  xmlSortAttributesByKey: boolean;
  /**
   * How to handle whitespaces in XML.
   *
   * @default "ignore"
   */
  xmlWhitespaceSensitivity: 'ignore' | 'preserve' | 'strict';
}

export type BuiltInParserName =
  | 'acorn'
  | 'angular'
  | 'babel'
  | 'babel-flow'
  | 'babel-ts'
  | 'css'
  | 'espree'
  | 'flow'
  | 'glimmer'
  | 'graphql'
  | 'html'
  | 'json5'
  | 'json'
  | 'json-stringify'
  | 'less'
  | 'lwc'
  | 'markdown'
  | 'mdx'
  | 'meriyah'
  | 'scss'
  | 'typescript'
  | 'vue'
  | 'xml'
  | 'yaml';

export type ExternalParserName = 'astro' | 'slidev';